requirements2.txt
absl-py                  1.4.0
asttokens                2.2.1
attrs                    23.1.0
backcall                 0.2.0
cachetools               5.3.1
certifi                  2023.7.22
cffi                     1.15.1
chardet                  5.2.0
charset-normalizer       3.2.0
colorama                 0.4.6
comm                     0.1.4
cryptography             41.0.3
dateparser               1.1.8
debugpy                  1.6.7.post1
decorator                5.1.1
distlib                  0.3.7
distro                   1.8.0
EasyProcess              1.1
entrypoint2              1.1
et-xmlfile               1.1.0
exceptiongroup           1.1.3
executing                1.2.0
ezdxf                    1.0.3
filelock                 3.12.2
fpdf                     1.7.2
google-api-core          2.11.1
google-auth              2.22.0
google-cloud-vision      3.4.4
googleapis-common-protos 1.60.0
grpcio                   1.57.0
grpcio-status            1.57.0
h11                      0.14.0
idna                     3.4
importlib-metadata       6.8.0
imutils                  0.5.4
invoice2data             0.4.4
ipykernel                6.25.1
ipython                  8.14.0
jedi                     0.19.0
jupyter_client           8.3.1
jupyter_core             5.3.1
lxml                     4.9.3
matplotlib-inline        0.1.6
mss                      9.0.1
nest-asyncio             1.5.7
numpy                    1.25.2
opencv-contrib-python    4.8.0.76
opencv-python            4.8.0.76
opencv-python-headless   4.8.0.76
openpyxl                 3.1.2
ortools                  9.7.2996
outcome                  1.2.0
packaging                23.1
pandas                   2.1.0
parso                    0.8.3
pdf2image                1.16.3
pdfminer.six             20221105
pdfplumber               0.10.2
pickleshare              0.7.5
Pillow                   10.0.0
pip                      23.2.1
platformdirs             3.10.0
pluggy                   1.2.0
prompt-toolkit           3.0.39
proto-plus               1.22.3
protobuf                 4.24.2
psutil                   5.9.5
pure-eval                0.2.2
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pycparser                2.21
pycryptodome             3.18.0
Pygments                 2.16.1
PyMuPDF                  1.23.3
PyMuPDFb                 1.23.3
pyparsing                3.1.1
PyPDF2                   3.0.1
pypdfium2                4.19.0
pyproject-api            1.5.4
pyscreenshot             3.1
PySocks                  1.7.1
pytesseract              0.3.10
python-dateutil          2.8.2
python-docx              0.8.11
pytz                     2023.3.post1
pywin32                  306
PyYAML                   6.0.1
pyzmq                    25.1.1
regex                    2023.8.8
requests                 2.31.0
rsa                      4.9
selenium                 4.11.2
setuptools               68.1.2
six                      1.16.0
sniffio                  1.3.0
sortedcontainers         2.4.0
stack-data               0.6.2
tabula-py                2.7.0
tomli                    2.0.1
tornado                  6.3.3
tox                      4.10.0
traitlets                5.9.0
trio                     0.22.2
trio-websocket           0.10.3
typing_extensions        4.7.1
tzdata                   2023.3
tzlocal                  5.0.1
urllib3                  1.26.16
virtualenv               20.24.3
wcwidth                  0.2.6
wheel                    0.41.2
wsproto                  1.2.0
zipp                     3.16.2